refactor: [017-FIX-WRONG-TERM] 상품 카테고리, 판매상태, 정렬기준 도메인에 대한 Production -> Product 오타 수정
