+++ src/main/java/com/myecommerce/MyECommerce/dto/production/ResponseSearchDetailProductionDto.java
@@ -1,7 +1,7 @@
 package com.myecommerce.MyECommerce.dto.production;
 
-import com.myecommerce.MyECommerce.type.ProductionCategoryType;
-import com.myecommerce.MyECommerce.type.ProductionSaleStatusType;
+import com.myecommerce.MyECommerce.type.ProductCategoryType;
+import com.myecommerce.MyECommerce.type.ProductSaleStatusType;
 import lombok.*;
 
 import java.time.LocalDateTime;
@@ -22,11 +22,11 @@ public class ResponseSearchDetailProductionDto {
 
     private String name;
 
-    private ProductionCategoryType category;
+    private ProductCategoryType category;
 
     private String description;
 
-    private ProductionSaleStatusType saleStatus;
+    private ProductSaleStatusType saleStatus;
 
     private Long createId;
