+++ src/main/java/com/myecommerce/MyECommerce/dto/production/ServiceProductionDto.java
@@ -1,7 +1,7 @@
 package com.myecommerce.MyECommerce.dto.production;
 
-import com.myecommerce.MyECommerce.type.ProductionCategoryType;
-import com.myecommerce.MyECommerce.type.ProductionSaleStatusType;
+import com.myecommerce.MyECommerce.type.ProductCategoryType;
+import com.myecommerce.MyECommerce.type.ProductSaleStatusType;
 import lombok.*;
 
 import java.util.List;
@@ -15,8 +15,8 @@ public class ServiceProductionDto {
     private Long id;
     private String code;
     private String name;
-    private ProductionCategoryType category;
+    private ProductCategoryType category;
     private String description;
-    private ProductionSaleStatusType saleStatus;
+    private ProductSaleStatusType saleStatus;
     private List<ServiceProductionOptionDto> options;
 }
