+++ src/main/java/com/myecommerce/MyECommerce/repository/product/ProductRepository.java
@@ -1,8 +1,8 @@
 package com.myecommerce.MyECommerce.repository.product;
 
 import com.myecommerce.MyECommerce.entity.product.Product;
-import com.myecommerce.MyECommerce.type.ProductionCategoryType;
-import com.myecommerce.MyECommerce.type.ProductionSaleStatusType;
+import com.myecommerce.MyECommerce.type.ProductCategoryType;
+import com.myecommerce.MyECommerce.type.ProductSaleStatusType;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -21,8 +21,8 @@ public interface ProductRepository extends JpaRepository<Product, Long> {
 
     // keyword를 포함하는 상품목록조회 - 최신등록순
     Page<Product> findByNameLikeAndSaleStatusAndCategoryOrderByCreateDt(
-            String name, ProductionSaleStatusType saleStatus,
-            ProductionCategoryType category, Pageable pageable);
+            String name, ProductSaleStatusType saleStatus,
+            ProductCategoryType category, Pageable pageable);
 
     // keyword를 포함하는 상품목록조회 - 낮은가격순
     @Query(value =
@@ -35,7 +35,7 @@ public interface ProductRepository extends JpaRepository<Product, Long> {
             " GROUP BY prd.id" +
             " ORDER BY MIN(option.price)")
     Page<Product> findByNameOrderByPrice(
-            String name, ProductionCategoryType category, Pageable pageable);
+            String name, ProductCategoryType category, Pageable pageable);
 
     // keyword를 포함하는 상품목록조회 - 높은가격순
     @Query(value =
@@ -48,7 +48,7 @@ public interface ProductRepository extends JpaRepository<Product, Long> {
             " GROUP BY prd.id" +
             " ORDER BY MIN(option.price) DESC")
     Page<Product> findByNameOrderByPriceDesc(
-            String name, ProductionCategoryType category, Pageable pageable);
+            String name, ProductCategoryType category, Pageable pageable);
 
     // keyword를 포함하는 상품목록조회 - 정확도순 (키워드 외 일치하지 않는 문자 개수로 판단)
     @Query(value =
@@ -59,5 +59,5 @@ public interface ProductRepository extends JpaRepository<Product, Long> {
             " AND prd.category = :category" +
             " ORDER BY (LENGTH(prd.name) - LENGTH(:name)) ")
     Page<Product> findByNameOrderByCalculatedAccuracyDesc(
-           String name, ProductionCategoryType category, Pageable pageable);
+           String name, ProductCategoryType category, Pageable pageable);
 }
