+++ src/main/java/com/myecommerce/MyECommerce/type/ProductOrderByStdType.java
@@ -4,7 +4,7 @@ import lombok.AllArgsConstructor;
 
 // 상품정렬기준 도메인
 @AllArgsConstructor
-public enum ProductionOrderByStdType {
+public enum ProductOrderByStdType {
     ORDER_BY_ACCURACY("정확도순"),
     ORDER_BY_REGISTRATION("등록순"),
     ORDER_BY_HIGHEST_PRICE("가격높은순"),
