+++ src/main/java/com/myecommerce/MyECommerce/type/ProductSaleStatusType.java
@@ -4,7 +4,7 @@ import lombok.AllArgsConstructor;
 
 // 상품판매상태 도메인
 @AllArgsConstructor
-public enum ProductionSaleStatusType {
+public enum ProductSaleStatusType {
     ON_SALE("판매중"),
     DISCONTINUED("판매중단"),
     DELETION("삭제")
